+++ src/DevilDaggersInfo.Core.Replay/Utils/SpanWrite.cs
@@ -4,10 +4,7 @@ public static class SpanWrite
 {
 	public static bool TryWriteChar(Span<char> destination, ref int charsWritten, char value)
 	{
-		if (destination.IsEmpty)
-			return false;
-
-		if (charsWritten + 1 >= destination.Length)
+		if (charsWritten + 1 > destination.Length)
 			return false;
 
 		destination[charsWritten++] = value;
@@ -16,10 +13,7 @@ public static class SpanWrite
 
 	public static bool TryWriteString(Span<char> destination, ref int charsWritten, string value)
 	{
-		if (destination.IsEmpty)
-			return false;
-
-		if (charsWritten + value.Length >= destination.Length)
+		if (charsWritten + value.Length > destination.Length)
 			return false;
 
 		value.AsSpan().CopyTo(destination[charsWritten..]);
@@ -27,16 +21,42 @@ public static class SpanWrite
 		return true;
 	}
 
+	public static bool TryWriteByte(Span<byte> utf8Destination, ref int bytesWritten, byte value)
+	{
+		if (bytesWritten + 1 > utf8Destination.Length)
+			return false;
+
+		utf8Destination[bytesWritten++] = value;
+		return true;
+	}
+
+	public static bool TryWriteBytes(Span<byte> utf8Destination, ref int bytesWritten, ReadOnlySpan<byte> value)
+	{
+		if (bytesWritten + value.Length > utf8Destination.Length)
+			return false;
+
+		value.CopyTo(utf8Destination[bytesWritten..]);
+		bytesWritten += value.Length;
+		return true;
+	}
+
 	public static bool TryWrite<T>(Span<char> destination, ref int charsWritten, T value, ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
 		where T : ISpanFormattable
 	{
-		if (destination.IsEmpty)
-			return false;
-
 		if (!value.TryFormat(destination[charsWritten..], out int charsWrittenValue, format, provider))
 			return false;
 
 		charsWritten += charsWrittenValue;
+		return true;
+	}
+
+	public static bool TryWrite<T>(Span<byte> utf8Destination, ref int bytesWritten, T value, ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
+		where T : IUtf8SpanFormattable
+	{
+		if (!value.TryFormat(utf8Destination[bytesWritten..], out int bytesWrittenValue, format, provider))
+			return false;
+
+		bytesWritten += bytesWrittenValue;
 		return true;
 	}
 }